Rename the generator file.
